feat(Demo): try to load custom config before initializing app
Default custom config file: lappconfig.json (the same directory as index.html)

lappconfig.json has the same format as lappconfigdefault.json,
except that all fields are allowed to be omitted for the default value.

This file does not need to present;
if the loading fails, the default config is used.

* add config _ConfigFileName for specifying the path to the custom config file
    * The prefix underscore `_` indicates that it is a nonstandard option
* src/main.ts:
    * window.onload callback:
        * load the config before initializing the app (LAppDelegate)
        * `await` the config loading to ensure that the config is ready